+++ lib/domain/model/contact.dart
@@ -89,9 +89,13 @@ class ChatContactId extends NewType<String> {
 /// Position of a [ChatContact] in a favorites list of the authenticated
 /// [MyUser].
 @HiveType(typeId: ModelTypeId.chatContactPosition)
-class ChatContactPosition extends NewType<double> {
+class ChatContactPosition extends NewType<double>
+    implements Comparable<ChatContactPosition> {
   const ChatContactPosition(double val) : super(val);
 
   factory ChatContactPosition.parse(String val) =>
       ChatContactPosition(double.parse(val));
+
+  @override
+  int compareTo(ChatContactPosition other) => val.compareTo(other.val);
 }