+++ lib/ui/page/home/tab/contacts/controller.dart
@@ -16,7 +16,6 @@
 
 import 'dart:async';
 
-import 'package:collection/collection.dart';
 import 'package:get/get.dart';
 
 import '/domain/model/chat.dart';
@@ -35,8 +34,8 @@ import '/domain/repository/user.dart';
 import '/domain/service/call.dart';
 import '/domain/service/contact.dart';
 import '/l10n/l10n.dart';
-import '/provider/gql/exceptions.dart' show UpdateChatContactNameException;
-import '/ui/widget/text_field.dart';
+import '/provider/gql/exceptions.dart'
+    show FavoriteChatContactException, UnfavoriteChatContactException;
 import '/util/message_popup.dart';
 import '/util/obs/obs.dart';
 
@@ -51,14 +50,8 @@ class ContactsTabController extends GetxController {
     this._settingsRepository,
   );
 
-  /// [TextFieldState] of a [ChatContact.name].
-  late TextFieldState contactName;
-
   /// Returns current reactive [ChatContact]s list.
   final RxList<RxChatContact> contacts = RxList<RxChatContact>();
-
-  /// [ChatContactId] of a [ChatContact] to rename.
-  final Rx<ChatContactId?> contactToChangeNameOf = Rx<ChatContactId?>(null);
 
   /// [Chat] repository used to create a dialog [Chat].
   final AbstractChatRepository _chatRepository;
@@ -85,9 +78,11 @@ class ContactsTabController extends GetxController {
   /// [StreamSubscription]s to the [contacts] updates.
   StreamSubscription? _contactsSubscription;
 
-  /// Returns the current reactive favorite [ChatContact]s map.
-  RxMap<ChatContactId, RxChatContact> get favorites =>
-      _contactService.favorites;
+  /// [StreamSubscription]s to the [favorites] updates.
+  StreamSubscription? _favoritesSubscription;
+
+  /// Reactive list of sorted [ChatContact]s.
+  late final RxList<RxChatContact> favorites;
 
   /// Indicates whether [ContactService] is ready to be used.
   RxBool get contactsReady => _contactService.isReady;
@@ -98,67 +93,10 @@ class ContactsTabController extends GetxController {
       _settingsRepository.applicationSettings.value?.sortContactsByName ?? true;
 
   @override
-  void onInit() {
-    contacts.addAll(_contactService.contacts.values);
-    _sortContacts();
-
-    contactName = TextFieldState(
-      onChanged: (s) async {
-        s.error.value = null;
-
-        RxChatContact? contact = contacts.firstWhereOrNull(
-                (e) => e.contact.value.id == contactToChangeNameOf.value) ??
-            favorites.values.firstWhereOrNull(
-                (e) => e.contact.value.id == contactToChangeNameOf.value);
-        if (contact == null) return;
-
-        if (contact.contact.value.name.val == s.text) {
-          contactToChangeNameOf.value = null;
-          return;
-        }
-
-        UserName? name;
-
-        try {
-          name = UserName(s.text);
-        } on FormatException catch (_) {
-          s.error.value = 'err_incorrect_input'.l10n;
-        }
-
-        if (s.error.value == null) {
-          try {
-            s.status.value = RxStatus.loading();
-            s.editable.value = false;
-
-            await _contactService.changeContactName(
-              contactToChangeNameOf.value!,
-              name!,
-            );
-
-            s.clear();
-            contactToChangeNameOf.value = null;
-          } on UpdateChatContactNameException catch (e) {
-            s.error.value = e.toMessage();
-          } catch (e) {
-            s.error.value = e.toString();
-            rethrow;
-          } finally {
-            s.editable.value = true;
-            s.status.value = RxStatus.empty();
-          }
-        }
-      },
-      onSubmitted: (s) {
-        RxChatContact? contact = contacts.firstWhereOrNull(
-                (e) => e.contact.value.id == contactToChangeNameOf.value) ??
-            favorites.values.firstWhereOrNull(
-                (e) => e.contact.value.id == contactToChangeNameOf.value);
-        if (contact?.contact.value.name.val == s.text) {
-          contactToChangeNameOf.value = null;
-          return;
-        }
-      },
-    );
+  void onInit() {    contacts.addAll(_contactService.contacts.values);
+  _sortContacts();
+    favorites = RxList(_contactService.favorites.values.toList());
+    _sortFavorites();
 
     _initUsersUpdates();
 
@@ -172,6 +110,7 @@ class ContactsTabController extends GetxController {
     }
 
     _contactsSubscription?.cancel();
+    _favoritesSubscription?.cancel();
     _rxUserWorkers.forEach((_, v) => v.dispose());
     _userWorkers.forEach((_, v) => v.dispose());
     super.onClose();
@@ -303,12 +242,67 @@ class ContactsTabController extends GetxController {
           break;
 
         case OperationKind.updated:
-          // No-op.
+          // no-op
+          break;
+      }
+    });
+
+    favorites.forEach(listen);
+
+    _favoritesSubscription = _contactService.favorites.changes.listen((e) {
+      switch (e.op) {
+        case OperationKind.added:
+          favorites.add(e.value!);
+          _sortFavorites();
+          listen(e.value!);
+          break;
+
+        case OperationKind.removed:
+          e.value?.user.value?.stopUpdates();
+          _userWorkers.remove(e.key)?.dispose();
+          favorites.removeWhere((c) => c.contact.value.id == e.key);
+          break;
+
+        case OperationKind.updated:
+          _sortFavorites();
           break;
       }
 
       _sortContacts();
     });
+  }
+
+  /// Marks the specified [ChatContact] identified by its [id] as favorited.
+  Future<void> favoriteContact(ChatContactId id) async {
+    try {
+      await _contactService.favoriteChatContact(id);
+    } on FavoriteChatContactException catch (e) {
+      MessagePopup.error(e);
+    } catch (e) {
+      MessagePopup.error(e);
+      rethrow;
+    }
+  }
+
+  /// Removes the specified [ChatContact] identified by its [id] from the
+  /// favorites.
+  Future<void> unfavoriteContact(ChatContactId id) async {
+    try {
+      await _contactService.unfavoriteChatContact(id);
+    } on UnfavoriteChatContactException catch (e) {
+      MessagePopup.error(e);
+    } catch (e) {
+      MessagePopup.error(e);
+      rethrow;
+    }
+  }
+
+  /// Sorts the [favorites] by the [ChatContact.favoritePosition].
+  void _sortFavorites() {
+    favorites.sort(
+      (a, b) => a.contact.value.favoritePosition!
+          .compareTo(b.contact.value.favoritePosition!),
+    );
   }
 }