+++ CHANGELOG.md
@@ -25,6 +25,8 @@ All user visible changes to this project will be documented in this file. This p
         - Quick status changing menu. ([#204], [#203])
     - Media panel:
         - Participants redialing. ([#241], [#233])
+    - Contacts tab:
+        - Favorite contacts. ([#237], [#223])
 
 ### Changed
 
@@ -84,8 +86,10 @@ All user visible changes to this project will be documented in this file. This p
 [#221]: /../../pull/221
 [#226]: /../../issues/226
 [#232]: /../../issues/232
+[#223]: /../../issues/223
 [#233]: /../../issues/233
 [#235]: /../../pull/235
+[#237]: /../../pull/237
 [#241]: /../../pull/241
 [#245]: /../../pull/245
 [#246]: /../../pull/246